upload page added and verification of resources link with user

+++ app/Http/Controllers/urlpageController.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Models\URL;
@@ -10,24 +10,24 @@ use App\Models\Visitor;
 class urlpageController extends Controller
 {
 public function index(){
-    // $urls = URL::all();
-    //Auth()->logout();
      // Ensure the user is authenticated
         $userID = auth()->user()->id;
+         Log::info($userID);
         $urls= URL::where("user_id", $userID)->get();
-        // $user=auth()->user();
-        // $urls=$user->urls;
         return view('urls.index', compact('urls'));
 
 }
 
 public function view($id){
     $url = URL::findOrFail($id);
+    if($url->user_id != auth()->user()->id){
+        abort(404);
+    }
     return view("urls.view",compact("url"));
-}
+    }
 public function create(){
     return view("urls.create");
-}
+    }
 public function store(Request $request){
     $request ->validate(
         [
@@ -46,6 +46,11 @@ public function store(Request $request){
 }
 public function edit($id){
     $url = URL::findOrFail($id);
+    //checks if user is accessing its own data or not
+    if($url->user_id != auth()->user()->id){
+        abort(404);
+    }
+    Log::info($url);
   //  dd($url);
     return view('urls.edit',compact('url'));
 }
@@ -56,6 +61,9 @@ public function update(Request $request, $id){
     ] );
 
    $url = URL::findOrFail($id);
+   if($url->user_id != auth()->user()->id){
+    abort(404);
+    }
    $url->update(
    ['original_url'=> $request->url]
    );
@@ -64,6 +72,9 @@ public function update(Request $request, $id){
 }
 public function destroy(Request $request ,$id){
     $url = URL::findOrFail($id);
+    if($url->user_id != auth()->user()->id){
+        abort(404);
+    }
     $url->delete();
     $request->session()-> flash('success','URL was deleted successfully');
     return redirect()->action([urlpageController::class,'index']);
